+++ src/backend/src/routes/index.js
@@ -2,11 +2,12 @@ const express = require('express');
 const authRoutes = require('./auth.routes');
 const userRoutes = require('./user.routes');
 const vehicleRoutes = require('./vehicle.routes');
-const routeRoutes   = require('./route.routes');
+const routeRoutes = require('./route.routes');
 const driverVerifRoutes = require('./driverVerification.routes');
 const bookingRoutes = require('./booking.routes');
 const notificationRoutes = require('./notification.routes')
 const mapRoutes = require('./maps.routes')
+const reviewRoutes = require('./review.routes')
 
 const router = express.Router();
 
@@ -17,6 +18,7 @@ router.use('/routes', routeRoutes);
 router.use('/driver-verifications', driverVerifRoutes);
 router.use('/bookings', bookingRoutes);
 router.use('/notifications', notificationRoutes);
+router.use('/reviews', reviewRoutes);
 router.use('/api/maps', mapRoutes);
 
 module.exports = router;